mm-improve add vfs

+++ src/fs/fs.h
@@ -75,5 +75,4 @@ struct dirent {
 // Function declarations
 struct inode* namei(char *path);
 struct inode* nameiat(struct inode *base, char *path);
-struct inode* createat(struct inode *dp, char *name, short type, short major, short minor);
-
+struct inode* createat(struct inode *dp, char *name, short type, int major, int minor);
